+++ tests/test_tei_transformer.py
@@ -4,7 +4,7 @@ import unittest
 from lxml import etree
 
 from tei_transform.observer import TeiNamespaceObserver
-from tei_transform.revision_desc_change import RevisionDescChange
+from tei_transform.parse_config import RevisionDescChange
 from tei_transform.tei_transformer import TeiTransformer
 from tei_transform.xml_tree_iterator import XMLTreeIterator
